services/test_service/requirements.txt
# --- Web framework ---
fastapi==0.110.1
mangum==0.17.0  # AWS Lambda adapter for FastAPI

# --- AWS SDK ---
boto3==1.34.89
botocore==1.34.89

# --- PostgreSQL Client ---
psycopg2-binary==2.9.9

# --- Logging / Utility ---
python-dotenv==1.0.1  # optional but useful if testing locally
